+++ src/trace_syscalls.h
@@ -5,8 +5,8 @@
 
 
 /* -- Function prototypes -- */
-const char *get_syscall_name(int syscall_nr);
-void print_syscall_args(pid_t pid, int syscall_nr);
+const char *get_syscall_name(long syscall_nr);
+void print_syscall_args(pid_t pid, long syscall_nr);
 
 
 #endif /* TRACE_SYSCALLS_H */
